display weekview calendar on reserve page
