+++ src/components/calendar/resources.ts
@@ -1,0 +1,27 @@
+const resourcesData = [
+  {
+    text: "初學者",
+    id: 1,
+    color: "#105861",
+  },
+  {
+    text: "一般練習者",
+    id: 2,
+    color: "#c76035",
+  },
+  {
+    text: "進階練習者",
+    id: 3,
+    color: "#cb4641",
+  },
+];
+
+const resources = [
+  {
+    fieldName: "Level",
+    title: "Level",
+    instances: resourcesData,
+  },
+];
+
+export default resources;
