+++ src/pages/_app.tsx
@@ -4,7 +4,7 @@ import { createGlobalStyle } from "styled-components";
 import { Reset } from "styled-reset";
 import Header from "../components/header";
 import Footer from "../components/footer";
-import { AuthContextProvider } from "../context/authContext";
+import { AuthContextProvider } from "../contexts/authContext";
 
 const GlobalStyle = createGlobalStyle`
 *,
